changed the type to size_t

+++ Algorithm.h
@@ -8,21 +8,21 @@ using namespace std;
 
 
 class Algorithm {
-    unsigned int *text;
-    unsigned int *pattern;
-    long unsigned int len_t;
-    long unsigned int len_p;
-    vector<long unsigned int> a;
-    vector<long unsigned int> automat;
-    vector<long unsigned int> lastSeenT;
+    size_t *text;
+    size_t *pattern;
+    size_t len_t;
+    size_t len_p;
+    vector<size_t> a;
+    vector<size_t> automat;
+    vector<size_t> lastSeenT;
     void createA(int size_ab);
     void createAutomat();
-    bool compareAutomat(unsigned int j , unsigned int i);
-    bool compareAlgorithm(unsigned int j , unsigned int i);
+    bool compareAutomat(size_t j , size_t i);
+    bool compareAlgorithm(size_t j , size_t i);
 
 public:
-    Algorithm(unsigned int *text, long unsigned int len_t, unsigned int *pattern, long unsigned int len_p, int size_ab);
-    list <long unsigned int> runAlgorithm();
+    Algorithm(size_t *text, size_t len_t, size_t *pattern, size_t len_p, int size_ab);
+    list <size_t> runAlgorithm();
 
 };
