+++ Naive.cpp
@@ -1,18 +1,18 @@
 #include "Naive.h"
 
-list <long unsigned int> naiveInfAB(unsigned int *text, long unsigned int len_t, unsigned int *pattern, long unsigned int len_p) {
-    list <long unsigned int> matches; //list where there where matches
+list <size_t> naiveInfAB(size_t *text, size_t len_t, size_t *pattern, size_t len_p) {
+    list <size_t> matches; //list where there where matches
     //create maps to match letters in text and pattern. need 2 to make sure different letters do not get the same match
-    unordered_map<unsigned int, long unsigned int> p_match;
-    unordered_map<unsigned int, bool> t_match;
+    unordered_map<size_t, size_t> p_match;
+    unordered_map<size_t, bool> t_match;
     //for every position in the text find p-match
-    for (long unsigned int i = 0; i <= len_t-len_p; i++) {
+    for (size_t i = 0; i <= len_t-len_p; i++) {
         //clear maps for new use
         p_match.clear();
         t_match.clear();
         bool match = true;
         //check for each letter if is matched
-        for (long unsigned int j = 0; j < len_p; j++) {
+        for (size_t j = 0; j < len_p; j++) {
             if (naiveCompareInfAB(p_match, t_match, pattern[j], text[i + j]) == false) {
                 match = false;
                 break;
@@ -25,7 +25,7 @@ list <long unsigned int> naiveInfAB(unsigned int *text, long unsigned int len_t,
     return matches;
 }
 
-bool naiveCompareInfAB(unordered_map<unsigned int, long unsigned int> &p_match, unordered_map<unsigned int, bool> &t_match, unsigned int p, unsigned int t) {
+bool naiveCompareInfAB(unordered_map<size_t, size_t> &p_match, unordered_map<size_t, bool> &t_match, size_t p, size_t t) {
     auto p_key = p_match.find(p); //check if letter p in map
     auto t_key = t_match.find(t); //check if letter t in map
     if (p_key == p_match.end() && t_key == t_match.end()) { //both letters did not appear before
@@ -39,20 +39,20 @@ bool naiveCompareInfAB(unordered_map<unsigned int, long unsigned int> &p_match, 
     }
 }
 
-list <long unsigned int> naive(unsigned int *text, long unsigned int len_t, unsigned int *pattern, long unsigned int len_p, int size_ab) {
-    list <long unsigned int> matches; //list where there where matches
+list <size_t> naive(size_t *text, size_t len_t, size_t *pattern, size_t len_p, int size_ab) {
+    list <size_t> matches; //list where there where matches
     //create arrays to match letters in text and pattern. need 2 to make sure different letters do not get the same match
-    std::vector<unsigned int> p_match(size_ab), t_match(size_ab);
+    std::vector<size_t> p_match(size_ab), t_match(size_ab);
 
     //for every position in the text find p-match
-    for (long unsigned int i = 0; i <= len_t-len_p; i++) {
+    for (size_t i = 0; i <= len_t-len_p; i++) {
         //create arrays to match letters in text and pattern. need 2 to make sure different letters do not get the same match
         t_match.assign(size_ab, size_ab+1);
         p_match.assign(size_ab, size_ab+1);
 
         bool match = true;
         //check for each letter if is matched
-        for (long unsigned int j = 0; j < len_p; j++) {
+        for (size_t j = 0; j < len_p; j++) {
             if (naiveCompare(p_match.data(), t_match.data(), pattern[j], text[i + j], size_ab) == false) {
                 match = false;
                 break;
@@ -65,7 +65,7 @@ list <long unsigned int> naive(unsigned int *text, long unsigned int len_t, unsi
     return matches;
 }
 
-bool naiveCompare(unsigned int *p_match, unsigned int *t_match, unsigned int p, unsigned int t, int size_ab) {
+bool naiveCompare(size_t *p_match, size_t *t_match, size_t p, size_t t, int size_ab) {
     if (p_match[p] == size_ab+1 && t_match[t] == size_ab+1) {
         p_match[p] = t;
         t_match[t] = p;
